Instructions/part list Update (#215)
* Updated Assembly Instructions

Updated instructions to account for the changes made to the base of V4.  Also added some pictures of the various hanging options.

* Updated Assembly Instructions

Edits to #3 and #4 to add more specifics.

* Update Instructions

Added details to the credentials section of the software instructions.
Added a link to an Adafruit wishlist in the parts sections and an explanation on how to load that wishlist into your cart.