Render Docs in a Webpage (#218)
Adds a step to build the Markdown files as a webpage which includes a nicer table of contents, loads faster, and looks better on mobile. This is implemented using MKDocs Material.

It also removes all the V1->V3 docs and Frog files as those are not really useful any longer.